backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1

# Telegram Bot
python-telegram-bot==21.7

# S3 Storage
boto3==1.35.44
minio==7.2.8

# Scheduled Tasks
apscheduler==3.10.4

# Utilities
pydantic==2.9.2
pydantic-settings==2.6.0
python-dateutil==2.9.0.post0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Development
black==24.10.0
ruff==0.7.2
